+++ api/src/main/java/org/jmisb/api/video/DemuxerUtils.java
@@ -1,14 +1,19 @@
 package org.jmisb.api.video;
 
-import org.bytedeco.javacpp.avcodec;
-import org.bytedeco.javacpp.avformat;
-import org.bytedeco.javacpp.avutil;
+import org.bytedeco.ffmpeg.avcodec.AVPacket;
+import org.bytedeco.ffmpeg.avformat.AVFormatContext;
+import org.bytedeco.ffmpeg.avformat.AVStream;
+import org.bytedeco.ffmpeg.avutil.AVRational;
 import org.jmisb.core.video.FfmpegUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static org.bytedeco.javacpp.avformat.*;
-import static org.bytedeco.javacpp.avutil.AVERROR_EOF;
+import static org.bytedeco.ffmpeg.global.avformat.AVSEEK_FLAG_BACKWARD;
+import static org.bytedeco.ffmpeg.global.avformat.av_guess_frame_rate;
+import static org.bytedeco.ffmpeg.global.avformat.av_read_frame;
+import static org.bytedeco.ffmpeg.global.avformat.avformat_flush;
+import static org.bytedeco.ffmpeg.global.avformat.avformat_seek_file;
+import static org.bytedeco.ffmpeg.global.avutil.AVERROR_EOF;
 
 /**
  * Utility methods for demuxing
@@ -27,16 +32,16 @@ class DemuxerUtils
      * @param inputStream The input stream
      * @return A new VideoDecodeThread, or null if one could not be created
      */
-    static VideoDecodeThread createVideoDecodeThread(int videoStreamIndex, avformat.AVFormatContext avFormatContext, VideoInput inputStream)
+    static VideoDecodeThread createVideoDecodeThread(int videoStreamIndex, AVFormatContext avFormatContext, VideoInput inputStream)
     {
         if (videoStreamIndex >= 0)
         {
-            avformat.AVStream videoStream = FfmpegUtils.getVideoStream(avFormatContext);
+            AVStream videoStream = FfmpegUtils.getVideoStream(avFormatContext);
             VideoDecodeThread videoDecodeThread = new VideoDecodeThread(inputStream, videoStream);
 
             if (logger.isDebugEnabled())
             {
-                avutil.AVRational frameRate = av_guess_frame_rate(avFormatContext, videoStream, null);
+                AVRational frameRate = av_guess_frame_rate(avFormatContext, videoStream, null);
                 logger.debug("Detected frame rate = " + frameRate.num() + " / " + frameRate.den());
             }
 
@@ -53,7 +58,7 @@ class DemuxerUtils
      * @param packet The packet
      * @return True if a packet was read, false otherwise
      */
-    static boolean readPacket(avformat.AVFormatContext avFormatContext, avcodec.AVPacket packet)
+    static boolean readPacket(AVFormatContext avFormatContext, AVPacket packet)
     {
         int ret;
         if ((ret = av_read_frame(avFormatContext, packet)) < 0)
@@ -81,7 +86,7 @@ class DemuxerUtils
      * @param avFormatContext The format context
      * @param seekPosition The position
      */
-    static void seek(avformat.AVFormatContext avFormatContext, double seekPosition)
+    static void seek(AVFormatContext avFormatContext, double seekPosition)
     {
         long microseconds = Math.round(seekPosition * 1000000);
         if (logger.isDebugEnabled())