+++ api/src/main/java/org/jmisb/api/video/MetadataDecodeThread.java
@@ -1,7 +1,8 @@
 package org.jmisb.api.video;
 
-import org.bytedeco.javacpp.avcodec;
-import org.bytedeco.javacpp.avformat;
+import org.bytedeco.ffmpeg.avcodec.AVCodecContext;
+import org.bytedeco.ffmpeg.avcodec.AVPacket;
+import org.bytedeco.ffmpeg.avformat.AVStream;
 import org.jmisb.api.common.KlvParseException;
 import org.jmisb.api.klv.KlvParser;
 import org.jmisb.api.klv.IMisbMessage;
@@ -9,17 +10,16 @@ import org.jmisb.core.video.FfmpegUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.nio.ByteBuffer;
-import java.nio.ByteOrder;
 import java.util.List;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingDeque;
 import java.util.concurrent.TimeUnit;
 
-import static org.bytedeco.javacpp.avcodec.av_packet_clone;
-import static org.bytedeco.javacpp.avutil.av_q2d;
+import static org.bytedeco.ffmpeg.global.avcodec.av_packet_clone;
+import static org.bytedeco.ffmpeg.global.avcodec.avcodec_alloc_context3;
+import static org.bytedeco.ffmpeg.global.avcodec.avcodec_free_context;
+import static org.bytedeco.ffmpeg.global.avcodec.avcodec_parameters_to_context;
+import static org.bytedeco.ffmpeg.global.avutil.av_q2d;
 
 /**
  * Metadata decoding thread
@@ -31,8 +31,8 @@ class MetadataDecodeThread extends ProcessingThread
     private static Logger logger = LoggerFactory.getLogger(MetadataDecodeThread.class);
     private final static int INPUT_QUEUE_SIZE = 100;
     private final VideoInput inputStream;
-    private final avformat.AVStream dataStream;
-    private BlockingQueue<avcodec.AVPacket> packetQueue = new LinkedBlockingDeque<>(INPUT_QUEUE_SIZE);
+    private final AVStream dataStream;
+    private BlockingQueue<AVPacket> packetQueue = new LinkedBlockingDeque<>(INPUT_QUEUE_SIZE);
 
     /**
      * Constructor
@@ -40,7 +40,7 @@ class MetadataDecodeThread extends ProcessingThread
      * @param inputStream The {@link VideoInput}
      * @param dataStream The metadata stream
      */
-    MetadataDecodeThread(VideoInput inputStream, avformat.AVStream dataStream)
+    MetadataDecodeThread(VideoInput inputStream, AVStream dataStream)
     {
         this.inputStream = inputStream;
         this.dataStream = dataStream;
@@ -53,7 +53,7 @@ class MetadataDecodeThread extends ProcessingThread
      * @param packet The packet to queue
      * @return True if the packet was queued, false if the queue is currently full
      */
-    public boolean enqueue(avcodec.AVPacket packet)
+    public boolean enqueue(AVPacket packet)
     {
         try
         {
@@ -75,13 +75,13 @@ class MetadataDecodeThread extends ProcessingThread
     {
         Thread.currentThread().setName("MetadataDecodeThread - " + inputStream.getUrl());
 
-        avcodec.AVCodecContext codecContext = avcodec.avcodec_alloc_context3(null);
+        AVCodecContext codecContext = avcodec_alloc_context3(null);
         int ret;
-        if ((ret = avcodec.avcodec_parameters_to_context(codecContext, dataStream.codecpar())) < 0)
+        if ((ret = avcodec_parameters_to_context(codecContext, dataStream.codecpar())) < 0)
         {
             logger.error("Couldn't create AVCodecContext for stream " + dataStream.index() +
                         " codec: " + dataStream.codecpar().codec_tag() + " error " + FfmpegUtils.formatError(ret));
-            avcodec.avcodec_free_context(codecContext);
+            avcodec_free_context(codecContext);
             return;
         }
 
@@ -96,7 +96,7 @@ class MetadataDecodeThread extends ProcessingThread
 
             try
             {
-                avcodec.AVPacket packet = packetQueue.poll(10, TimeUnit.MILLISECONDS);
+                AVPacket packet = packetQueue.poll(10, TimeUnit.MILLISECONDS);
                 if (packet != null)
                 {
                     double pts = packet.pts() * av_q2d(dataStream.time_base());
@@ -133,6 +133,6 @@ class MetadataDecodeThread extends ProcessingThread
         if (logger.isDebugEnabled())
             logger.debug("Data stream decoder exiting");
 
-        avcodec.avcodec_free_context(codecContext);
+        avcodec_free_context(codecContext);
     }
 }