+++ api/src/main/java/org/jmisb/api/video/VideoInput.java
@@ -1,6 +1,6 @@
 package org.jmisb.api.video;
 
-import org.bytedeco.javacpp.avformat;
+import org.bytedeco.ffmpeg.avformat.AVFormatContext;
 import org.jmisb.core.video.FfmpegUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -17,8 +17,8 @@ import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.TimeUnit;
 
-import static org.bytedeco.javacpp.avformat.avformat_close_input;
-import static org.bytedeco.javacpp.avformat.avformat_free_context;
+import static org.bytedeco.ffmpeg.global.avformat.avformat_close_input;
+import static org.bytedeco.ffmpeg.global.avformat.avformat_free_context;
 import static org.jmisb.core.video.TimingUtils.shortWait;
 
 /**
@@ -42,7 +42,7 @@ public abstract class VideoInput implements IVideoInput
     private final static int QUEUE_SIZE = 100;
 
     String url;
-    avformat.AVFormatContext formatContext;
+    AVFormatContext formatContext;
 
     @Override
     public abstract void open(String url) throws IOException;
@@ -53,8 +53,20 @@ public abstract class VideoInput implements IVideoInput
     @Override
     public abstract void close();
 
+    /**
+     * Insert a sleep to control playback rate prior to notifying clients
+     *
+     * @param pts PTS of the video frame to be delivered
+     */
     protected abstract void delayVideo(double pts);
-    protected abstract void delayMetadata(double pts);
+
+    /**
+     * Insert a sleep to control playback rate prior to notifying clients
+     *
+     * @param pts PTS of the metadata frame to be delivered
+     * @throws InterruptedException If the thread is interrupted while waiting
+     */
+    protected abstract void delayMetadata(double pts) throws InterruptedException;
 
     @Override
     public String getUrl()
@@ -149,14 +161,10 @@ public abstract class VideoInput implements IVideoInput
      */
     void startNotifiers(boolean startPaused)
     {
-        videoNotifier = new VideoNotifier();
-        if (startPaused)
-            videoNotifier.pauseOutput();
+        videoNotifier = new VideoNotifier(startPaused);
         videoNotifier.start();
 
-        metadataNotifier = new MetadataNotifier();
-        if (startPaused)
-            metadataNotifier.pauseOutput();
+        metadataNotifier = new MetadataNotifier(startPaused);
         metadataNotifier.start();
     }
 
@@ -195,9 +203,14 @@ public abstract class VideoInput implements IVideoInput
      */
     protected class VideoNotifier extends Thread
     {
-        private boolean shutdown = false;
+        private volatile boolean shutdown = false;
         private boolean paused = false;
         private boolean getOneFrame = false;
+
+        VideoNotifier(boolean paused)
+        {
+            this.paused = paused;
+        }
 
         @Override
         public void run()
@@ -238,6 +251,7 @@ public abstract class VideoInput implements IVideoInput
         void shutdown()
         {
             shutdown = true;
+            interrupt();
         }
 
         protected void pauseOutput()
@@ -261,8 +275,13 @@ public abstract class VideoInput implements IVideoInput
      */
     protected class MetadataNotifier extends Thread
     {
-        private boolean shutdown = false;
+        private volatile boolean shutdown = false;
         private boolean paused = false;
+
+        MetadataNotifier(boolean paused)
+        {
+            this.paused = paused;
+        }
 
         @Override
         public void run()
@@ -299,6 +318,7 @@ public abstract class VideoInput implements IVideoInput
         void shutdown()
         {
             shutdown = true;
+            interrupt();
         }
 
         protected void pauseOutput()