+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### May 11, 2019 version 1.7.0
+ * Update to JavaCPP Presets 1.5 / FFmpeg 4.1.3
+ * Option to enable or disable KLV stream in output
+ * Metadata framing performance optimizations
+ * KLV parser wraps unsupported message types and returns them in a
+   RawMisbMessage
+ * Metadata usage example fixed
+ 
 ### September 14, 2018 version 1.6.0
  * Updated FFmpeg from 3.4.2 to 4.0.1
  * Added API methods to get PES information as a JSON string